requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
numpy==1.25.0
pandas==2.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
QualtricsAPI==0.5.8
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.3
